Fix molten fluids not existing since #4655
